+++ model/vowpalWabbit/dffml_model_vowpalWabbit/vw_base.py
@@ -210,11 +210,12 @@ class VWContext(ModelContext):
                 X = vw_data
             for x in X:
                 self.parent.clf.learn(x)
+        self.is_trained = True
 
     async def predict(
         self, sources: SourcesContext
     ) -> AsyncIterator[Tuple[Record, Any, float]]:
-        if not os.path.isfile(self.parent.model_filename):
+        if not self.is_trained:
             raise ModelNotTrained("Train model before prediction.")
         importance, tag, base = None, None, None
         if self.parent.config.importance:
@@ -321,6 +322,7 @@ class VWModel(Model):
         self.features = self.applicable_features(self.config.features)
         self._features_hash = self._feature_predict_hash()
         self.clf = None
+        self.is_trained = self.model_filename.exists()
 
     def applicable_features(self, features):
         usable = []